Обновление защиты
Обновил защиту на удаление и редактирование рецепта

+++ CookingRecipesApi/CookingRecipesApi/Controllers/RecipeController.cs
@@ -8,6 +8,7 @@ using Domain.Recipes.Entities;
 using FluentValidation;
 using FluentValidation.Results;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CookingRecipesApi.Controllers;
@@ -67,10 +68,17 @@ public class RecipeController : ControllerBase
     [Route( "update/{recipeId}" )]
     public async Task<IActionResult> UpdateRecipe( [FromForm] RecipeDto recipeDto, [FromRoute] int recipeId )
     {
+        int authorId = int.Parse( User.GetUserId() );
+
+        bool hasAccess = await _recipeService.HasAccessToRecipe( recipeId, authorId );
+
+        if ( !hasAccess )
+        {
+            return StatusCode( 403, new ErrorResponse( "Нет доступа" ) );
+        }
+
         try
         {
-            int authorId = int.Parse( User.GetUserId() );
-
             await _recipeService.UpdateRecipe( recipeDto.ToDomain( authorId ), recipeId, _appEnvironment.WebRootPath );
 
             await _unitOfWork.Save();
@@ -87,6 +95,15 @@ public class RecipeController : ControllerBase
     [Route( "delete/{recipeId}" )]
     public async Task<IActionResult> RemoveRecipe( [FromRoute] int recipeId )
     {
+        int authorId = int.Parse( User.GetUserId() );
+
+        bool hasAccess = await _recipeService.HasAccessToRecipe( recipeId, authorId );
+
+        if ( !hasAccess )
+        {
+            return StatusCode( 403, new ErrorResponse( "Нет доступа" ) );
+        }
+
         try
         {
             await _recipeService.RemoveRecipe( recipeId, _appEnvironment.WebRootPath );
